+++ service/src/test/java/uk/gov/hmcts/reform/fpl/model/CaseDataTest.java
@@ -10,7 +10,7 @@ import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.jackson.JacksonAutoConfiguration;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
-import uk.gov.hmcts.reform.fpl.model.generatedorder.FurtherDirections;
+import uk.gov.hmcts.reform.fpl.model.order.generated.FurtherDirections;
 
 import java.util.Optional;
