+++ service/src/test/java/uk/gov/hmcts/reform/fpl/service/GeneratedOrderServiceTest.java
@@ -27,9 +27,9 @@ import uk.gov.hmcts.reform.fpl.model.common.Element;
 import uk.gov.hmcts.reform.fpl.model.common.JudgeAndLegalAdvisor;
 import uk.gov.hmcts.reform.fpl.model.emergencyprotectionorder.EPOChildren;
 import uk.gov.hmcts.reform.fpl.model.emergencyprotectionorder.EPOPhrase;
-import uk.gov.hmcts.reform.fpl.model.generatedorder.FurtherDirections;
-import uk.gov.hmcts.reform.fpl.model.generatedorder.GeneratedOrder;
-import uk.gov.hmcts.reform.fpl.model.generatedorder.InterimEndDate;
+import uk.gov.hmcts.reform.fpl.model.order.generated.FurtherDirections;
+import uk.gov.hmcts.reform.fpl.model.order.generated.GeneratedOrder;
+import uk.gov.hmcts.reform.fpl.model.order.generated.InterimEndDate;
 import uk.gov.hmcts.reform.fpl.service.config.LookupTestConfig;
 import uk.gov.hmcts.reform.fpl.service.time.Time;
 import uk.gov.hmcts.reform.fpl.utils.FixedTimeConfiguration;
